+++ angular-workspace/projects/trimble-oss/modus-angular-components/src/lib/stencil-generated/components.ts
@@ -246,6 +246,34 @@ export class ModusChip {
   }
 }
 
+import type { ModusDataTableSortEvent as IModusDataTableModusDataTableSortEvent } from '@trimble-oss/modus-web-components';
+export declare interface ModusDataTable extends Components.ModusDataTable {
+  /**
+   * An event that fires on column sort. 
+   */
+  sort: EventEmitter<CustomEvent<IModusDataTableModusDataTableSortEvent>>;
+
+}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['columns', 'data', 'size', 'sortOptions']
+})
+@Component({
+  selector: 'modus-data-table',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['columns', 'data', 'size', 'sortOptions']
+})
+export class ModusDataTable {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+    proxyOutputs(this, this.el, ['sort']);
+  }
+}
+
 
 export declare interface ModusDropdown extends Components.ModusDropdown {
   /**
@@ -859,13 +887,13 @@ export declare interface ModusTreeView extends Components.ModusTreeView {}
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['checkboxSelection', 'disabled', 'multiCheckboxSelection', 'multiSelection', 'size']
+  inputs: ['checkboxSelection', 'checkedItems', 'expandedItems', 'multiCheckboxSelection', 'multiSelection', 'selectedItems', 'size']
 })
 @Component({
   selector: 'modus-tree-view',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
-  inputs: ['checkboxSelection', 'disabled', 'multiCheckboxSelection', 'multiSelection', 'size']
+  inputs: ['checkboxSelection', 'checkedItems', 'expandedItems', 'multiCheckboxSelection', 'multiSelection', 'selectedItems', 'size']
 })
 export class ModusTreeView {
   protected el: HTMLElement;
@@ -878,6 +906,10 @@ export class ModusTreeView {
 
 export declare interface ModusTreeViewItem extends Components.ModusTreeViewItem {
   /**
+   * An event that fires on tree item checkbox click 
+   */
+  checkboxClick: EventEmitter<CustomEvent<boolean>>;
+  /**
    * An event that fires on tree item click 
    */
   itemClick: EventEmitter<CustomEvent<boolean>>;
@@ -885,16 +917,13 @@ export declare interface ModusTreeViewItem extends Components.ModusTreeViewItem 
    * An event that fires on tree item expand/collapse 
    */
   itemExpandToggle: EventEmitter<CustomEvent<boolean>>;
-  /**
-   * An event that fires on tree item checkbox click 
-   */
-  checkboxClick: EventEmitter<CustomEvent<boolean>>;
 
 }
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['checked', 'disabled', 'expanded', 'indeterminate', 'label', 'nodeId', 'selected']
+  inputs: ['checked', 'disabled', 'expanded', 'indeterminate', 'label', 'nodeId', 'selected'],
+  methods: ['focusItem']
 })
 @Component({
   selector: 'modus-tree-view-item',
@@ -907,6 +936,6 @@ export class ModusTreeViewItem {
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
     this.el = r.nativeElement;
-    proxyOutputs(this, this.el, ['itemClick', 'itemExpandToggle', 'checkboxClick']);
+    proxyOutputs(this, this.el, ['checkboxClick', 'itemClick', 'itemExpandToggle']);
   }
 }