+++ stencil-workspace/src/components.d.ts
@@ -6,10 +6,12 @@
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { Crumb } from "./components/modus-breadcrumb/modus-breadcrumb";
+import { ModusDataTableSortEvent, ModusTableSortOptions, TCell, TColumn, TRow } from "./components/modus-data-table/modus-data-table.models";
 import { App } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
 import { App as App1 } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
 import { RadioButton } from "./components/modus-radio-group/modus-radio-button";
 import { Tab } from "./components/modus-tabs/modus-tabs";
+import { TreeViewItemOptions } from "./components/modus-content-tree/types";
 export namespace Components {
     interface ModusAccordion {
         /**
@@ -180,6 +182,18 @@ export namespace Components {
           * (optional) The chip's value.
          */
         "value": string;
+    }
+    interface ModusDataTable {
+        "columns": string[] | TColumn[];
+        "data": TCell[][] | TRow[];
+        /**
+          * The size of the table.
+         */
+        "size"?: 'condensed' | 'standard';
+        /**
+          * Options for data table column sort.
+         */
+        "sortOptions"?: ModusTableSortOptions;
     }
     interface ModusDropdown {
         /**
@@ -715,9 +729,13 @@ export namespace Components {
          */
         "checkboxSelection": boolean;
         /**
-          * (optional) Disables each tree item
+          * (optional) Checked tree items by default
          */
-        "disabled": boolean;
+        "checkedItems": string[];
+        /**
+          * (optional) Expanded tree items by default
+         */
+        "expandedItems": string[];
         /**
           * (optional) Enables multiple checkbox selection
          */
@@ -726,6 +744,10 @@ export namespace Components {
           * (optional) Enables multiple tree items selection
          */
         "multiSelection": boolean;
+        /**
+          * (optional) Selected tree items by default
+         */
+        "selectedItems": string[];
         /**
           * (optional) The default size of all tree items
          */
@@ -744,6 +766,7 @@ export namespace Components {
           * (optional) Expanded state of the tree item
          */
         "expanded": boolean;
+        "focusItem": () => Promise<void>;
         /**
           * (optional) Checkbox indeterminate state of the tree item
          */
@@ -787,6 +810,10 @@ export interface ModusCheckboxCustomEvent<T> extends CustomEvent<T> {
 export interface ModusChipCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLModusChipElement;
+}
+export interface ModusDataTableCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLModusDataTableElement;
 }
 export interface ModusDropdownCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -906,6 +933,12 @@ declare global {
     var HTMLModusChipElement: {
         prototype: HTMLModusChipElement;
         new (): HTMLModusChipElement;
+    };
+    interface HTMLModusDataTableElement extends Components.ModusDataTable, HTMLStencilElement {
+    }
+    var HTMLModusDataTableElement: {
+        prototype: HTMLModusDataTableElement;
+        new (): HTMLModusDataTableElement;
     };
     interface HTMLModusDropdownElement extends Components.ModusDropdown, HTMLStencilElement {
     }
@@ -1067,6 +1100,7 @@ declare global {
         "modus-card": HTMLModusCardElement;
         "modus-checkbox": HTMLModusCheckboxElement;
         "modus-chip": HTMLModusChipElement;
+        "modus-data-table": HTMLModusDataTableElement;
         "modus-dropdown": HTMLModusDropdownElement;
         "modus-file-dropzone": HTMLModusFileDropzoneElement;
         "modus-list": HTMLModusListElement;
@@ -1296,6 +1330,22 @@ declare namespace LocalJSX {
           * (optional) The chip's value.
          */
         "value"?: string;
+    }
+    interface ModusDataTable {
+        "columns"?: string[] | TColumn[];
+        "data"?: TCell[][] | TRow[];
+        /**
+          * An event that fires on column sort.
+         */
+        "onSort"?: (event: ModusDataTableCustomEvent<ModusDataTableSortEvent>) => void;
+        /**
+          * The size of the table.
+         */
+        "size"?: 'condensed' | 'standard';
+        /**
+          * Options for data table column sort.
+         */
+        "sortOptions"?: ModusTableSortOptions;
     }
     interface ModusDropdown {
         /**
@@ -1885,9 +1935,13 @@ declare namespace LocalJSX {
          */
         "checkboxSelection"?: boolean;
         /**
-          * (optional) Disables each tree item
+          * (optional) Checked tree items by default
          */
-        "disabled"?: boolean;
+        "checkedItems"?: string[];
+        /**
+          * (optional) Expanded tree items by default
+         */
+        "expandedItems"?: string[];
         /**
           * (optional) Enables multiple checkbox selection
          */
@@ -1896,6 +1950,10 @@ declare namespace LocalJSX {
           * (optional) Enables multiple tree items selection
          */
         "multiSelection"?: boolean;
+        /**
+          * (optional) Selected tree items by default
+         */
+        "selectedItems"?: string[];
         /**
           * (optional) The default size of all tree items
          */
@@ -1956,6 +2014,7 @@ declare namespace LocalJSX {
         "modus-card": ModusCard;
         "modus-checkbox": ModusCheckbox;
         "modus-chip": ModusChip;
+        "modus-data-table": ModusDataTable;
         "modus-dropdown": ModusDropdown;
         "modus-file-dropzone": ModusFileDropzone;
         "modus-list": ModusList;
@@ -1996,6 +2055,7 @@ declare module "@stencil/core" {
             "modus-card": LocalJSX.ModusCard & JSXBase.HTMLAttributes<HTMLModusCardElement>;
             "modus-checkbox": LocalJSX.ModusCheckbox & JSXBase.HTMLAttributes<HTMLModusCheckboxElement>;
             "modus-chip": LocalJSX.ModusChip & JSXBase.HTMLAttributes<HTMLModusChipElement>;
+            "modus-data-table": LocalJSX.ModusDataTable & JSXBase.HTMLAttributes<HTMLModusDataTableElement>;
             "modus-dropdown": LocalJSX.ModusDropdown & JSXBase.HTMLAttributes<HTMLModusDropdownElement>;
             "modus-file-dropzone": LocalJSX.ModusFileDropzone & JSXBase.HTMLAttributes<HTMLModusFileDropzoneElement>;
             "modus-list": LocalJSX.ModusList & JSXBase.HTMLAttributes<HTMLModusListElement>;